+++ src/runtime/Evaluator.ts
@@ -1080,12 +1080,25 @@ export default class Evaluator {
         }
     }
 
+    /** Keep track of whether we're stepping to detect accidental infinite recursion. */
+    private stepping = false;
+
     /**
      * Advance one step in execution.
      * Return any exceptions or the final value when there's nothing left to execute.
      * Otherwise, return undefined, as a signal that we're still stepping.
      **/
     step(): void {
+        // Check if we're stepping. If so, it means something in the step below is causing the Evaluator
+        // to start stepping again, which should never happen.
+        if (this.stepping) {
+            console.error(
+                'Already stepping, likely infinite recursion detected.',
+            );
+            console.trace();
+        }
+        this.stepping = true;
+
         // Get the evaluation to step.
         const evaluation = this.getCurrentEvaluation();
 
@@ -1140,6 +1153,17 @@ export default class Evaluator {
         if (!this.isInPast()) {
             this.#totalStepCount++;
             this.#stepCount = this.#stepIndex;
+        }
+
+        // Done stepping, reset the flag.
+        this.stepping = false;
+
+        // If there are queued changes, react to them.
+        if (this.isDone() && this.queuedChanges.length > 0) {
+            const changes = this.queuedChanges;
+            this.queuedChanges = [];
+            // React to the changes.
+            this.evaluate(changes);
         }
     }
 
@@ -1555,7 +1579,16 @@ export default class Evaluator {
         }
     }
 
+    // Changes we haven't processed yet, because we're not done evaluating.
+    private queuedChanges: StreamValue[] = [];
+
     evaluate(changed: StreamValue[]) {
+        // If we're starting a new evaluation, but not done, add the changes to the queue and let end() evaluate them after we're done.
+        if (!this.isDone()) {
+            this.queuedChanges.push(...changed);
+            return;
+        }
+
         // One or more streams changed! Let's find out what expressions to update.
         for (const stream of changed) {
             const streamNode = stream.creator;